requirements.txt
flask
scikit-learn
pandas
numpy
matplotlib
seaborn
mysql-connector-python
joblib
flask-cors
